% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{prep_data_for_plot}
\alias{prep_data_for_plot}
\title{Title}
\usage{
prep_data_for_plot(
  .data,
  focus_var,
  key_name = "key",
  values_name = "Proportion (\%)",
  rounding = 3,
  include_summary_line = FALSE,
  recoding = NULL,
  type = "bar",
  bar_column = `1`,
  key_order = NULL
)
}
\arguments{
\item{.data}{}

\item{focus_var}{}

\item{key_name}{}

\item{values_name}{}

\item{rounding}{}

\item{include_summary_line}{}

\item{recoding}{}

\item{type}{}

\item{bar_column}{}

\item{key_order}{}
}
\value{

}
\description{
Title
}
