% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarehhs.R
\docType{package}
\name{rarehhs-package}
\alias{rarehhs}
\alias{rarehhs-package}
\title{rarehhs: PKG_TITLE}
\description{
PKG_DESC.
}
\author{
\strong{Maintainer}: AUTHOR_FIRST AUTHOR_LAST \email{AUTHOR@MAIL.COM}

}
