% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_side_performance_indicators.R
\name{sidebarPerIndServer}
\alias{sidebarPerIndServer}
\title{sidebar performance indicators Server Function}
\usage{
sidebarPerIndServer(id, state)
}
\description{
sidebar performance indicators Server Function
}
